feat: bump guest platform MEM_SIZE to 2^32 bytes
Raises the guest runtime memory bound (allocator ceiling, is_guest_memory,
RVR guarded bounds) from 2^29 to the full 2^32-byte circuit capacity, so the
platform and circuit memory sizes coincide again.

- Elf::decode max_mem widened u32 -> u64: at 2^32 the previous
  `MEM_SIZE as u32` arguments would wrap to 0. All call sites updated.
- RVR needs no C-side changes: MEMORY_MASK is emitted as a u64 literal and
  OPENVM_MEM_SIZE derives from it, so compiled execution now permits the
  full range; test_high_address_mem is un-gated from non-rvr builds and
  passes under RVR.
- out_of_bound_mem_access's wild address moves 8 bytes past 4 GiB (its old
  address, 8 bytes past 512 MiB, is now valid memory); the protected-rvr
  trap test still passes.
- Guests built with the prebuilt std toolchain keep the allocator ceiling
  baked into the openvm rust fork's PAL until a new toolchain release;
  no_std guests pick up the new ceiling from openvm-platform directly.

+++ crates/cli/src/commands/build.rs
@@ -397,7 +397,7 @@ pub fn build(build_args: &BuildArgs, cargo_args: &BuildCargoArgs) -> Result<Path
         let transpiler = app_config.app_vm_config.transpiler();
         let data = read(elf_path)
             .with_context(|| format!("failed to read ELF at {}", elf_path.display()))?;
-        let elf = Elf::decode(&data, MEM_SIZE as u32)
+        let elf = Elf::decode(&data, MEM_SIZE as u64)
             .with_context(|| format!("failed to decode ELF for target '{}'", target.name))?;
         let exe = VmExe::from_elf(elf, transpiler)
             .with_context(|| format!("failed to transpile target '{}'", target.name))?;
