+++ crates/toolchain/tests/tests/rv64_transpiler_tests.rs
@@ -20,7 +20,7 @@ use test_case::test_case;
 fn get_elf(elf_path: impl AsRef<Path>) -> Result<Elf> {
     let dir = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
     let data = read(dir.join(elf_path))?;
-    let elf = Elf::decode(&data, MEM_SIZE as u32)?;
+    let elf = Elf::decode(&data, MEM_SIZE as u64)?;
     Ok(elf)
 }
